Switched to official mongo driver

+++ event_store.go
@@ -4,5 +4,5 @@ type EventStore interface {
 	Append(events *EventStream) error
 	GetEventsFor(streamName StreamName, id string) (*EventStream, error)
 	FromVersion(streamName StreamName, id string, version int) (*EventStream, error)
-	CountEventsFor(streamName StreamName, id string) (int, error)
+	CountEventsFor(streamName StreamName, id string) (int64, error)
 }
